requirement.txt
pygame-ce==2.5.7
customtkinter
pandas
matplotlib
seaborn